# openapi.yaml
openapi: 3.1.0
info:
  title: База
  version: 0.1.0
paths:
  /auth/jwt/login:
    post:
      tags:
        - auth
      summary: Auth:Jwt.Login
      operationId: auth_jwt_login_auth_jwt_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Body_auth_jwt_login_auth_jwt_login_post"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BearerResponse"
              example:
                access_token: >-
                  eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoiOTIyMWZmYzktNjQwZi00MzcyLTg2ZDMtY2U2NDJjYmE1NjAzIiwiYXVkIjoiZmFzdGFwaS11c2VyczphdXRoIiwiZXhwIjoxNTcxNTA0MTkzfQ.M10bjOe45I5Ncu_uXvOmVV8QxnL-nZfcH96U90JaocI
                token_type: bearer
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
              examples:
                LOGIN_BAD_CREDENTIALS:
                  summary: Bad credentials or the user is inactive.
                  value:
                    detail: LOGIN_BAD_CREDENTIALS
                LOGIN_USER_NOT_VERIFIED:
                  summary: The user is not verified.
                  value:
                    detail: LOGIN_USER_NOT_VERIFIED
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /auth/jwt/logout:
    post:
      tags:
        - auth
      summary: Auth:Jwt.Logout
      operationId: auth_jwt_logout_auth_jwt_logout_post
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "401":
          description: Missing token or inactive user.
      security:
        - OAuth2PasswordBearer: []
  /auth/register:
    post:
      tags:
        - auth
      summary: Register:Register
      operationId: register_register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
        required: true
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
              examples:
                REGISTER_USER_ALREADY_EXISTS:
                  summary: A user with this email already exists.
                  value:
                    detail: REGISTER_USER_ALREADY_EXISTS
                REGISTER_INVALID_PASSWORD:
                  summary: Password validation failed.
                  value:
                    detail:
                      code: REGISTER_INVALID_PASSWORD
                      reason: Password should beat least 3 characters
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /users/me:
    get:
      tags:
        - users
      summary: Info Me
      operationId: info_me_users_me_get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
      security:
        - OAuth2PasswordBearer: []
    patch:
      tags:
        - users
      summary: Update User Me
      operationId: update_user_me_users_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateMe"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - OAuth2PasswordBearer: []
  /users/{user_id}:
    patch:
      tags:
        - users
      summary: Update Users
      operationId: update_users_users__user_id__patch
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    get:
      tags:
        - users
      summary: Info User
      operationId: info_user_users__user_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - users
      summary: Deactivate User
      operationId: deactivate_user_users__user_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRead"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /task/{task_id}/user/{executor_id}/:
    post:
      tags:
        - task
      summary: Select User For A Task
      description: Выбор исполнителя для задачи заказчиком
      operationId: select_user_for_a_task_task__task_id__user__executor_id___post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            title: Task Id
        - name: executor_id
          in: path
          required: true
          schema:
            type: integer
            title: Executor Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDB"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /task/create/:
    post:
      tags:
        - task
      summary: Create Task
      description: Создание задачи заказчиком
      operationId: create_task_task_create__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDB"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - OAuth2PasswordBearer: []
  /task/open/:
    get:
      tags:
        - task
      summary: Get Not Busy Tasks
      description: Получение свободных задач, на которые еще не выбрано исполнитель
      operationId: get_not_busy_tasks_task_open__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            description: Page number
            default: 1
            title: Page
          description: Page number
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            description: Page size
            default: 50
            title: Size
          description: Page size
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_TaskDB_"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /task/all/:
    get:
      tags:
        - task
      summary: All
      description: Получение всех задач
      operationId: all_task_all__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            description: Page number
            default: 1
            title: Page
          description: Page number
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            description: Page size
            default: 50
            title: Size
          description: Page size
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_TaskDB_"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /task/my_tasks/:
    get:
      tags:
        - task
      summary: My
      description: Задачи исполнителя
      operationId: my_task_my_tasks__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            description: Page number
            default: 1
            title: Page
          description: Page number
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            description: Page size
            default: 50
            title: Size
          description: Page size
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_TaskDB_"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /task/my_responsed_tasks/:
    get:
      tags:
        - task
      summary: My Responsed Tasks
      description: Задачи, на которые откликнулся исполнитель
      operationId: my_responsed_tasks_task_my_responsed_tasks__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            description: Page number
            default: 1
            title: Page
          description: Page number
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            description: Page size
            default: 50
            title: Size
          description: Page size
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_TaskDB_"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /task/done_task_executor/{task_id}/:
    post:
      tags:
        - task
      summary: Done Task By Ex
      operationId: done_task_by_ex_task_done_task_executor__task_id___post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            title: Task Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDB"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /task/done_task_customer/{task_id}/:
    post:
      tags:
        - task
      summary: Done Task By Cus
      operationId: done_task_by_cus_task_done_task_customer__task_id___post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            title: Task Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDB"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /task/delete/{task_id}/:
    delete:
      tags:
        - task
      summary: Cancel Task
      operationId: cancel_task_task_delete__task_id___delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            title: Task Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDB"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /bid/who_responsed_by_task/{task_id}:
    get:
      tags:
        - bid
      summary: Who Responsed By Task
      operationId: who_responsed_by_task_bid_who_responsed_by_task__task_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            title: Task Id
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            description: Page number
            default: 1
            title: Page
          description: Page number
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            description: Page size
            default: 50
            title: Size
          description: Page size
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page_BidWithExecutor_"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /bid/{task_id}/response/:
    post:
      tags:
        - bid
      summary: Response To A Task
      operationId: response_to_a_task_bid__task_id__response__post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            title: Task Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BidDB"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
components:
  schemas:
    BearerResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      type: object
      required:
        - access_token
        - token_type
      title: BearerResponse
    BidDB:
      properties:
        task_id:
          type: integer
          title: Task Id
        executor_id:
          type: integer
          title: Executor Id
      type: object
      required:
        - task_id
        - executor_id
      title: BidDB
    BidWithExecutor:
      properties:
        task_id:
          type: integer
          title: Task Id
        executor:
          $ref: "#/components/schemas/UserRead"
      type: object
      required:
        - task_id
        - executor
      title: BidWithExecutor
    Body_auth_jwt_login_auth_jwt_login_post:
      properties:
        grant_type:
          anyOf:
            - type: string
              pattern: ^password$
            - type: "null"
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          format: password
          title: Password
        scope:
          type: string
          title: Scope
          default: ""
        client_id:
          anyOf:
            - type: string
            - type: "null"
          title: Client Id
        client_secret:
          anyOf:
            - type: string
            - type: "null"
          format: password
          title: Client Secret
      type: object
      required:
        - username
        - password
      title: Body_auth_jwt_login_auth_jwt_login_post
    CreateTask:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        price:
          type: number
          title: Price
      type: object
      required:
        - title
        - description
        - price
      title: CreateTask
    ErrorModel:
      properties:
        detail:
          anyOf:
            - type: string
            - additionalProperties:
                type: string
              type: object
          title: Detail
      type: object
      required:
        - detail
      title: ErrorModel
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Page_BidWithExecutor_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/BidWithExecutor"
          type: array
          title: Items
        total:
          type: integer
          minimum: 0
          title: Total
        page:
          type: integer
          minimum: 1
          title: Page
        size:
          type: integer
          minimum: 1
          title: Size
        pages:
          type: integer
          minimum: 0
          title: Pages
      type: object
      required:
        - items
        - total
        - page
        - size
        - pages
      title: Page[BidWithExecutor]
    Page_TaskDB_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/TaskDB"
          type: array
          title: Items
        total:
          type: integer
          minimum: 0
          title: Total
        page:
          type: integer
          minimum: 1
          title: Page
        size:
          type: integer
          minimum: 1
          title: Size
        pages:
          type: integer
          minimum: 0
          title: Pages
      type: object
      required:
        - items
        - total
        - page
        - size
        - pages
      title: Page[TaskDB]
    TaskDB:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        price:
          type: number
          title: Price
        id:
          type: integer
          title: Id
        selected_executor_id:
          anyOf:
            - type: integer
            - type: "null"
          title: Selected Executor Id
        done_executor:
          type: boolean
          title: Done Executor
        done_customer:
          type: boolean
          title: Done Customer
      type: object
      required:
        - title
        - description
        - price
        - id
        - done_executor
        - done_customer
      title: TaskDB
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 32
          title: Username
        first_name:
          type: string
          maxLength: 32
          title: First Name
        last_name:
          type: string
          maxLength: 32
          title: Last Name
        role_id:
          type: integer
          title: Role Id
        password:
          type: string
          title: Password
      type: object
      required:
        - username
        - first_name
        - last_name
        - role_id
        - password
      title: UserCreate
    UserRead:
      properties:
        username:
          type: string
          maxLength: 32
          title: Username
        first_name:
          type: string
          maxLength: 32
          title: First Name
        last_name:
          type: string
          maxLength: 32
          title: Last Name
        role_id:
          type: integer
          title: Role Id
        id:
          type: integer
          title: Id
        balance:
          type: number
          title: Balance
      type: object
      required:
        - username
        - first_name
        - last_name
        - role_id
        - id
        - balance
      title: UserRead
    UserUpdate:
      properties:
        username:
          type: string
          maxLength: 32
          title: Username
        first_name:
          type: string
          maxLength: 32
          title: First Name
        last_name:
          type: string
          maxLength: 32
          title: Last Name
        role_id:
          anyOf:
            - type: integer
            - type: "null"
          title: Role Id
      type: object
      title: UserUpdate
    UserUpdateMe:
      properties:
        username:
          type: string
          maxLength: 32
          title: Username
        first_name:
          type: string
          maxLength: 32
          title: First Name
        last_name:
          type: string
          maxLength: 32
          title: Last Name
      type: object
      title: UserUpdateMe
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login